create new excess report file (not fully functional)
