+++ index.js
@@ -31,8 +31,10 @@ app.put('/users/:id', db.updateUser);
 app.delete('/users/:id', db.deleteUser);
 */
 app.get('/inventory_items', db.getInventoryItems);
-app.get('/excessReport',db.getExcessReport);
 
+
+const dbExcessReport = require('./excessReport');
+app.get('/excessReport',dbExcessReport.getExcessReport);
 
 
 app.post('/whatSalesTogether', getWhatSalesTogether);
